Add: Init method on instance variables
Initialize the game object's instance variables,
including the players with prompts to get the
player's names. This method includes the important
part of randomly assigning the symbols to each
player and announcing the game play as if the
Game class is a referee.

To break up an already long initialize method,
I decided to break off the logic for randomly
deciding which player would go first.