Add: Play method to run game logic with helpers
Main play method runs the game logic in a loop

There is a helper method to manage each turn

There is a helper method for the turn method to
change which player's turn it is at the end of
the previous player's turn. The reverse_turn
method seems to have a bug where it's not correctly
reversing turns.