Fix: Current player not reversing turn
I got trapped with an OOP issue. current_player
seemed to not be reversing turns. My method was
correct. I wasn't updating the instance variable
because I was missing the @ sign to update the
state of the variable.

Instead, I was updating a local variable. I fixed
it by adding the @.