Add: Logic for prompting when invalid moves
I want to add a message when the move is invalid
because the position is unavailable. There's a
disconnect between the Game class when the check
for valid_moves? happen from the GameBoard and
when to display the message to the Player.

This still needs to be tested and implemented.